massive refactoring

+++ src/lib/Replay.js
@@ -1,32 +1,51 @@
-const { Buffer } = require('safe-buffer');
-const fs = require('fs');
+const { Buffer } = require("safe-buffer");
+const fs = require("fs");
 const path = require("path");
-const spawn = require('child_process').spawn;
+const spawn = require("child_process").spawn;
+const { getReplay, createReplay } = require("./Database");
 
+/**
+ * Class for handling replays
+ */
 module.exports = class Replay {
     /**
-     * 
      * @param {String} filename Name of replay file
      * @param {Blob} base64Data Base64-encoded data of replay file
      */
     constructor(id, filename, base64Data) {
         this.id = id;
         this.filename = filename;
-        this.path = path.join(__dirname, '../../uploads/' + filename);
+        this.path = path.join(__dirname, "../../uploads/" + filename);
         this.base64Data = base64Data;
+        this.data = {};
     }
 
+    /**
+     * Adds folderName to path
+     * @param {String} folderName 
+     */
     addFolderToPath(folderName) {
-        this.path = path.join(__dirname, '../../uploads/' + folderName + "/" + this.filename);
+        this.path = path.join(
+            __dirname,
+            "../../uploads/" + folderName + "/" + this.filename
+        );
     }
 
+    /**
+     * Saves replay on disk
+     * @returns {Promise} of file save
+     */
     saveReplay() {
         return new Promise((resolve, reject) => {
-            fs.writeFile(this.path, Buffer.from(this.base64Data, 'base64'), (err) => {
-                if (err) reject(err);
-                resolve();
-            })
-        })
+            fs.writeFile(
+                this.path,
+                Buffer.from(this.base64Data, "base64"),
+                (err) => {
+                    if (err) reject(err);
+                    resolve();
+                }
+            );
+        });
     }
 
     /**
@@ -35,13 +54,17 @@ module.exports = class Replay {
      */
     saveAndExtractReplay() {
         return new Promise((resolve, reject) => {
-            fs.writeFile(this.path, Buffer.from(this.base64Data, 'base64'), (err) => {
-                if (err) reject({ status: 'failed', error: err });
-                this.extractInputsFromReplay()
-                .then(data => resolve(data))
-                .catch(e => reject({ status: 'failed', error: e }));
-            })
-        })
+            fs.writeFile(
+                this.path,
+                Buffer.from(this.base64Data, "base64"),
+                (err) => {
+                    if (err) reject({ status: "failed", error: err });
+                    this.extractInputsFromReplay()
+                        .then((data) => resolve(data))
+                        .catch((e) => reject({ status: "failed", error: e }));
+                }
+            );
+        });
     }
 
     /**
@@ -50,27 +73,55 @@ module.exports = class Replay {
      */
     extractInputsFromReplay() {
         return new Promise((resolve, reject) => {
-            const ls = spawn('python3', [path.join(__dirname, '../scripts/extract_replay.py'), this.path])
-        
+            const ls = spawn("python3", [
+                path.join(__dirname, "../scripts/extract_replay.py"),
+                this.path,
+            ]);
+
             let output = "";
 
-            ls.stdout.on('data', (data) => {
-                output += data
-            });
-            
-            ls.stderr.on('data', (data) => {
-                reject(data)
+            ls.stdout.on("data", (data) => {
+                output += data;
             });
 
-            ls.stdout.on('end', () => {
-                const replay = JSON.parse(output)[0]
-                replay.id = this.id
-                resolve([ replay ])
+            ls.stderr.on("data", (data) => {
+                reject(data);
             });
-        })
+
+            ls.stdout.on("end", () => {
+                const replay = JSON.parse(output)[0];
+                replay.id = this.id;
+                this.data = replay;
+                resolve([replay]);
+            });
+        });
     }
 
+    /**
+     * Creates replay in database if no replay with same replayUID exists
+     * @returns {Promise}
+     */
+    saveExtractedReplayToDatabaseIfNotExists() {
+        return new Promise((resolve, reject) => {
+            getReplay({ replayUID: this.data.replayUID })
+                .then((replay) => {
+                    if (replay !== null) resolve();
+                    else createReplay(this.data)
+                        .then(() => resolve())
+                        .catch((err) => reject(err));
+                })
+                .catch((err) => {
+                    reject(err);
+                });
+        });
+    }
+
+    /**
+     * Deletes file from disk if exists
+     */
     removeFileFromDisk() {
-        fs.rmSync(this.path);
+        if (fs.existsSync(this.path)) {
+            fs.rmSync(this.path);
+        }
     }
-}
+};
